common make_flat_edge: take 'ind' as a 'unsigned'
Pays down some cast noise from prior commits.

+++ lib/dotgen/dotsplines.c
@@ -77,8 +77,8 @@ static bool cl_vninside(Agraph_t *, Agnode_t *);
 static void completeregularpath(path *, Agedge_t *, Agedge_t *,
 				pathend_t *, pathend_t *, boxf *, int, int);
 static int edgecmp(const void *, const void *);
-static void make_flat_edge(graph_t *, spline_info_t *, path *, Agedge_t **, int,
-                           unsigned, int);
+static void make_flat_edge(graph_t *, spline_info_t *, path *, Agedge_t **,
+                           unsigned, unsigned, int);
 static void make_regular_edge(graph_t *g, spline_info_t *, path *, Agedge_t **,
                               int, unsigned, int);
 static boxf makeregularend(boxf, int, double);
@@ -451,7 +451,7 @@ static void dot_splines_(graph_t *g, int normalize) {
 	    }
 	}
 	else if (ND_rank(agtail(e0)) == ND_rank(aghead(e0))) {
-	    make_flat_edge(g, &sd, &P, edges, (int)ind, cnt, et);
+	    make_flat_edge(g, &sd, &P, edges, ind, cnt, et);
 	}
 	else
 	    make_regular_edge(g, &sd, &P, edges, (int)ind, cnt, et);
@@ -1550,7 +1550,7 @@ static void make_flat_bottom_edges(graph_t *g, spline_info_t *sp, path *P,
  *     = the rest - route along top
  */
 static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
-                           edge_t **edges, int ind, unsigned cnt, int et) {
+                           edge_t **edges, unsigned ind, unsigned cnt, int et) {
     node_t *tn, *hn;
     Agedgeinfo_t fwdedgei;
     Agedgepair_t fwdedge;
@@ -1570,7 +1570,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
 	e = &fwdedge.out;
     }
     for (unsigned i = 1; i < cnt; i++) {
-	if (ED_adjacent(edges[ind + (int)i])) {
+	if (ED_adjacent(edges[ind + i])) {
 	    isAdjacent = 1;
 	    break;
 	}
@@ -1579,7 +1579,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
      * so check them all.
      */
     if (isAdjacent) {
-	make_flat_adj_edges(g, edges, ind, cnt, e, et);
+	make_flat_adj_edges(g, edges, (int)ind, cnt, e, et);
 	return;
     }
     if (ED_label(e)) {  /* edges with labels aren't multi-edges */
@@ -1588,7 +1588,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
     }
 
     if (et == EDGETYPE_LINE) {
-	makeSimpleFlat(agtail(e), aghead(e), edges, ind, cnt, et);
+	makeSimpleFlat(agtail(e), aghead(e), edges, (int)ind, cnt, et);
 	return;
     }
 
@@ -1596,7 +1596,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
     hside = ED_head_port(e).side;
     if ((tside == BOTTOM && hside != TOP) ||
         (hside == BOTTOM && tside != TOP)) {
-	make_flat_bottom_edges(g, sp, P, edges, ind, cnt, e, et == EDGETYPE_SPLINE);
+	make_flat_bottom_edges(g, sp, P, edges, (int)ind, cnt, e, et == EDGETYPE_SPLINE);
 	return;
     }
 
@@ -1622,7 +1622,7 @@ static void make_flat_edge(graph_t *g, spline_info_t *sp, path *P,
 
     for (unsigned i = 0; i < cnt; i++) {
 	boxf b;
-	e = edges[ind + (int)i];
+	e = edges[ind + i];
 	size_t boxn = 0;
 
 	boxf boxes[3];
